+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Log critical exceptions to the application provided block via the new
+  configuration option ([#209](https://github.com/envato/event_sourcery/pull/209)):
+
+  ```ruby
+  config.on_event_processor_critical_error = proc do |exception, processor_name|
+    # report the death of this processor to an error reporting service like Rollbar.
+  end
+  ```
+
 ## [0.21.0] - 2018-07-02
 ### Added
 - Graceful shutdown interrupts poll-wait sleep for quicker quitting